+++ src/app/admin/subject/layout.tsx
@@ -1,7 +1,7 @@
 import {
   BlockBrowserNavigation,
   NavigationBlockerProvider,
-} from "./_components/navigation-block";
+} from "../../../components/subject/navigation-block";
 
 export default function DashboardLayout({
   children,
