+++ src/app/subject/[slug]/page.tsx
@@ -4,10 +4,10 @@ import Footer from "@/components/global/footer";
 import Navbar from "@/components/global/navbar";
 import { db } from "@/lib/firebase";
 import { type Subject } from "@/types";
-import SubjectBreadcrumb from "@/components/subjectHomepage/subject-breadcrumb";
-import SubjectSidebar from "@/components/subjectHomepage/subject-sidebar";
-import TableOfContents from "@/components/subjectHomepage/table-of-contents";
-import UnitAccordion from "@/components/subjectHomepage/unit-accordion";
+import SubjectBreadcrumb from "@/components/subject/subject-breadcrumb";
+import SubjectSidebar from "@/components/subject/subject-sidebar";
+import TableOfContents from "@/components/subject/table-of-contents";
+import UnitAccordion from "@/components/subject/unit-accordion";
 import { useEffect, useState } from "react";
 import { doc, getDoc } from "firebase/firestore";
 import { getUser } from "@/components/hooks/users";
