+++ src/components/subject/navigation-block.tsx
